+++ Synth.cpp
@@ -2,7 +2,7 @@
 #include "defines.h"
 #include "Synth.h"
 
-uint32_t Synth::run(uint8_t cycle)
+uint16_t Synth::run(uint8_t cycle)
 {
   return(this->last_module->run(cycle));
 }
